Onsyte first commit
Onsyte is a stand-in for FreeWave OnSight, the eventual
implementation of all of this learning.

This project plucks files out of other project directories and
installs them in various directories on the ZumLink.

It uses preinst to create the directories,
and postinst to change ownerships and permissions on the files.
(The chmod part doesn't work right now.)

It uses postrm to clean up the directory structure after removing
the files, etc.

Eventually it will install dependencies. Not there yet.

Eventually it will set up crontabs and stuff as well. Not there
yet.